+++ app/src/main/kotlin/com/simplemobiletools/gallery/activities/MainActivity.kt
@@ -402,12 +402,16 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
     }
 
     private fun increaseColumnCount() {
+        directories_vertical_fastscroller.measureRecyclerViewOnRedraw()
+        directories_horizontal_fastscroller.measureRecyclerViewOnRedraw()
         config.dirColumnCnt = ++(directories_grid.layoutManager as GridLayoutManager).spanCount
         invalidateOptionsMenu()
         directories_grid.adapter?.notifyDataSetChanged()
     }
 
     private fun reduceColumnCount() {
+        directories_vertical_fastscroller.measureRecyclerViewOnRedraw()
+        directories_horizontal_fastscroller.measureRecyclerViewOnRedraw()
         config.dirColumnCnt = --(directories_grid.layoutManager as GridLayoutManager).spanCount
         invalidateOptionsMenu()
         directories_grid.adapter?.notifyDataSetChanged()
@@ -472,6 +476,8 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
             inputStream = FileInputStream(File(path))
             outputStream = contentResolver.openOutputStream(output)
             inputStream.copyTo(outputStream)
+        } catch (e: SecurityException) {
+            showErrorToast(e)
         } catch (ignored: FileNotFoundException) {
         } finally {
             inputStream?.close()
@@ -523,9 +529,11 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
     }
 
     private fun gotDirectories(newDirs: ArrayList<Directory>, isFromCache: Boolean) {
-        Thread {
-            mLatestMediaId = getLatestMediaId()
-        }.start()
+        if (!isFromCache) {
+            Thread {
+                mLatestMediaId = getLatestMediaId()
+            }.start()
+        }
 
         val dirs = getSortedDirectories(newDirs)
         directories_refresh_layout.isRefreshing = false
@@ -533,6 +541,11 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
 
         directories_empty_text_label.beVisibleIf(dirs.isEmpty() && !isFromCache)
         directories_empty_text.beVisibleIf(dirs.isEmpty() && !isFromCache)
+        directories_grid.beVisibleIf(directories_empty_text_label.isGone())
+
+        val allowHorizontalScroll = config.scrollHorizontally && config.viewTypeFiles == VIEW_TYPE_GRID
+        directories_vertical_fastscroller.beVisibleIf(directories_grid.isVisible() && !allowHorizontalScroll)
+        directories_horizontal_fastscroller.beVisibleIf(directories_grid.isVisible() && allowHorizontalScroll)
 
         checkLastMediaChanged()
         if (dirs.hashCode() == mDirs.hashCode()) {
@@ -559,7 +572,8 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
         val currAdapter = directories_grid.adapter
         if (currAdapter == null) {
             initZoomListener()
-            DirectoryAdapter(this, mDirs, this, directories_grid, isPickIntent(intent) || isGetAnyContentIntent(intent)) {
+            val fastscroller = if (config.scrollHorizontally) directories_horizontal_fastscroller else directories_vertical_fastscroller
+            DirectoryAdapter(this, mDirs, this, directories_grid, isPickIntent(intent) || isGetAnyContentIntent(intent), fastscroller) {
                 itemClicked((it as Directory).path)
             }.apply {
                 setupZoomListener(mZoomListener)