+++ app/src/main/kotlin/com/simplemobiletools/gallery/adapters/MediaAdapter.kt
@@ -1,6 +1,8 @@
 package com.simplemobiletools.gallery.adapters
 
 import android.net.Uri
+import android.os.Handler
+import android.os.Looper
 import android.view.Menu
 import android.view.View
 import android.view.ViewGroup
@@ -12,27 +14,39 @@ import com.simplemobiletools.commons.dialogs.RenameItemDialog
 import com.simplemobiletools.commons.extensions.applyColorFilter
 import com.simplemobiletools.commons.extensions.beVisibleIf
 import com.simplemobiletools.commons.extensions.isActivityDestroyed
+import com.simplemobiletools.commons.views.FastScroller
 import com.simplemobiletools.commons.views.MyRecyclerView
 import com.simplemobiletools.gallery.R
 import com.simplemobiletools.gallery.dialogs.DeleteWithRememberDialog
 import com.simplemobiletools.gallery.extensions.*
 import com.simplemobiletools.gallery.helpers.VIEW_TYPE_LIST
 import com.simplemobiletools.gallery.models.Medium
-import kotlinx.android.synthetic.main.photo_video_item_list.view.*
+import kotlinx.android.synthetic.main.photo_video_item_grid.view.*
 import java.io.File
 import java.util.*
 
 class MediaAdapter(activity: BaseSimpleActivity, var media: MutableList<Medium>, val listener: MediaOperationsListener?, val isAGetIntent: Boolean,
-                   val allowMultiplePicks: Boolean, recyclerView: MyRecyclerView, itemClick: (Any) -> Unit) : MyRecyclerViewAdapter(activity, recyclerView, itemClick) {
+                   val allowMultiplePicks: Boolean, recyclerView: MyRecyclerView, fastScroller: FastScroller? = null,
+                   itemClick: (Any) -> Unit) : MyRecyclerViewAdapter(activity, recyclerView, fastScroller, itemClick) {
+
+    private val INSTANT_LOAD_DURATION = 2000L
+    private val IMAGE_LOAD_DELAY = 100L
 
     private val config = activity.config
     private val isListViewType = config.viewTypeFiles == VIEW_TYPE_LIST
     private var skipConfirmationDialog = false
+    private var visibleItemPaths = ArrayList<String>()
+    private var loadImageInstantly = false
+    private var delayHandler = Handler(Looper.getMainLooper())
 
     private var scrollHorizontally = config.scrollHorizontally
     private var animateGifs = config.animateGifs
     private var cropThumbnails = config.cropThumbnails
     private var displayFilenames = config.displayFileNames
+
+    init {
+        enableInstantLoad()
+    }
 
     override fun getActionMenuId() = R.menu.cab_media
 
@@ -51,6 +65,7 @@ class MediaAdapter(activity: BaseSimpleActivity, var media: MutableList<Medium>,
 
     override fun onBindViewHolder(holder: MyRecyclerViewAdapter.ViewHolder, position: Int) {
         val medium = media[position]
+        visibleItemPaths.add(medium.path)
         val view = holder.bindView(medium, !allowMultiplePicks) { itemView, layoutPosition ->
             setupView(itemView, medium)
         }
@@ -92,7 +107,9 @@ class MediaAdapter(activity: BaseSimpleActivity, var media: MutableList<Medium>,
     override fun onViewRecycled(holder: ViewHolder?) {
         super.onViewRecycled(holder)
         if (!activity.isActivityDestroyed()) {
-            Glide.with(activity).clear(holder?.itemView?.medium_thumbnail)
+            val itemView = holder?.itemView
+            visibleItemPaths.remove(itemView?.photo_name?.tag)
+            Glide.with(activity).clear(itemView?.medium_thumbnail!!)
         }
     }
 
@@ -230,12 +247,14 @@ class MediaAdapter(activity: BaseSimpleActivity, var media: MutableList<Medium>,
 
     fun updateMedia(newMedia: ArrayList<Medium>) {
         media = newMedia
+        enableInstantLoad()
         notifyDataSetChanged()
         finishActMode()
     }
 
     fun updateDisplayFilenames(displayFilenames: Boolean) {
         this.displayFilenames = displayFilenames
+        enableInstantLoad()
         notifyDataSetChanged()
     }
 
@@ -254,12 +273,32 @@ class MediaAdapter(activity: BaseSimpleActivity, var media: MutableList<Medium>,
         notifyDataSetChanged()
     }
 
+    private fun enableInstantLoad() {
+        loadImageInstantly = true
+        delayHandler.postDelayed({
+            loadImageInstantly = false
+        }, INSTANT_LOAD_DURATION)
+    }
+
     private fun setupView(view: View, medium: Medium) {
         view.apply {
             play_outline.beVisibleIf(medium.video)
             photo_name.beVisibleIf(displayFilenames || isListViewType)
             photo_name.text = medium.name
-            activity.loadImage(medium.path, medium_thumbnail, scrollHorizontally, animateGifs, cropThumbnails)
+            photo_name.tag = medium.path
+
+            if (loadImageInstantly) {
+                activity.loadImage(medium.path, medium_thumbnail, scrollHorizontally, animateGifs, cropThumbnails)
+            } else {
+                medium_thumbnail.setImageDrawable(null)
+                medium_thumbnail.isHorizontalScrolling = scrollHorizontally
+                delayHandler.postDelayed({
+                    val isVisible = visibleItemPaths.contains(medium.path)
+                    if (isVisible) {
+                        activity.loadImage(medium.path, medium_thumbnail, scrollHorizontally, animateGifs, cropThumbnails)
+                    }
+                }, IMAGE_LOAD_DELAY)
+            }
 
             if (isListViewType) {
                 photo_name.setTextColor(textColor)